Full Web Application Example (#1)
From the intro:

This is an example MVC web application rendered with razor pages. It is intended to show how you can integrate a Runly process into a web application workflow. After clicking the action button, a long-running, asynchronous process will be queued while providing real-time feedback to the user.