Refactor Runly.Testing classes (#1)
Rework the TestHost to have a similar API to JobHost. Instead of taking a ServiceCollection in the constructor, it now allows you to call ConfigureServices on it after it is created.

It also now registers logging via services.AddLogging in the TestHost since every test case was requiring that and someone writing a test case for their job would also have to do that to get internal classes to register correctly (e.g. ExecutionBase which requires a logger).

I also finally got GitHub Package Feeds working for this repo. So, for pre-release packages, we don't need to pollute our public nuget feed with 10,000 pre-release packages while we are testing things. It will now go to this repo's NuGet feed. I added a note in the README for how someone could consume the pre-release packages if they so desired.